Patched a possible exploit
The key combo stored in the config file now gets validated on config file load, since in some circumstances it will be passed as a parameter to another process through the command line, that would run with administrative/root privilages. This could allow an attacker to inject code into the command line with a malformed key combo stored in the config file, which would gain root access on the host computer.